Rename lib.progress to libprogress
