+++ generators/random.py
@@ -16,7 +16,7 @@ Copyright (C) 2020 BBaoVanC
 
 # Imports
 import random
-import lib.progress
+import libprogress
 
 
 # Generation method
@@ -45,11 +45,11 @@ def gen(count=1, debug=False, length=12):
             name = name + str(char)  # add the character to the name
         if debug:  # if we should output debug information
             # print progress bar for generating names
-            print(lib.progress.genbar(curprg=n-1, maxprg=count), end="\r")
+            print(libprogress.genbar(curprg=n-1, maxprg=count), end="\r")
         names.append(name)  # add the name to the name list
         n = n + 1  # increase counter
 
     if debug:  # if we should output debug information
         # print final progress bar
-        print(lib.progress.genbar(curprg=n-1, maxprg=count) + "...done")
+        print(libprogress.genfullbar(prg=count))
     return names  # return the name list
